// src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
// ...
export type FieldInputTypes = {
  readonly public: {
    readonly Project: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly repo: CodecTypes['pg/text@1']['input'];
      readonly defaultBranch: CodecTypes['pg/text@1']['input'];
      readonly testCommand: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Run: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly projectId: CodecTypes['pg/text@1']['input'];
      readonly status: 'pending' | 'running' | 'passed' | 'failed';
      readonly branch: CodecTypes['pg/text@1']['input'];
      readonly commitSha: CodecTypes['pg/text@1']['input'];
      readonly startedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly finishedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly duration: CodecTypes['pg/float8@1']['input'] | null;
      readonly passed: CodecTypes['pg/int4@1']['input'];
      readonly failed: CodecTypes['pg/int4@1']['input'];
      readonly skipped: CodecTypes['pg/int4@1']['input'];
      readonly total: CodecTypes['pg/int4@1']['input'];
      readonly files: CodecTypes['pg/json@1']['input'];
    };
  };
};